requirements.txt
flask
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-psycopg2
opentelemetry-exporter-otlp